refactor: revert xformCommGroupToCommonORder removal
